pyproject.toml
[project]
name = "chart-mcp-server"
version = "0.1.0"
description = "MCP Server for generating chart images from python code"
authors = [
    { name = "Bongjoo Seo", email = "qnnp@qnnp.net" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "matplotlib>=3.10.1",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.1",
]
